backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
roboflow==1.1.9
ultralytics==8.0.196
python-dotenv==1.0.0